Aplicando solid em camadas
